Change the semantics of readers
This changes the semantics of readers such that an offset must
_always_ be specified when creating a sub reader, and for stream
implementations, the stream is automatically wound forward to the
specified offset (raising an exception if this would imply a backwards
seek). This ensures that both the random access and stream readers
have the exact same API for creating sub readers and that code can
be written consistently for both.