Change the semantics of writers
This changes the semantics of writers such that an offset must
_always_ be specified when creating a sub writer, and for stream
implementations, the stream is automatically wound forward to the
specified offset (raising an exception if this would imply a backwards
seek). This ensures that both the random access and stream writers
have the exact same API for creating sub writers and that code can
be written consistently for both.

Additionally, this moves all exception creation into a single class
for consistent error messages.